Improved (ie, existent) error-checking of response from dyndns
Now we only update the lastip and expires files if we get a response
containing either the string "good" or the string "nochg". Anything else is
considered an error, and the update will be retried on the next run (since the
condition that caused the script to run the last time -- WAN IP change, or 28
days elapsed -- will still hold).

This change suggested itself after I got a "dnserr" response, but updated the
state on my end. Now dyndns thinks my hostname is due to expire 7 days before
my script does... this is not good.