Be more careful about IP addresses returned from checkip.dyndns.org
Occasionally we'll get a bogus result from calling checkip, and this results
in spurious, unneeded updates -- it looks as though our IP address has
changed, even though it hasn't.

The script has been changed to keep the result of last call to checkip (for
debugging) and to be more careful about parsing out the IP address, and to do
nothing if it didn't find an IP address.